+++ src/renderer/eval/PeopleList.tsx
@@ -1,9 +1,23 @@
-import { Button, Form, Input, List, Modal, Popconfirm, Spin } from 'antd';
-import 'antd/dist/antd.css';
-import { Link } from 'react-router-dom';
-import { PlusOutlined } from '@ant-design/icons';
-import { useState } from 'react';
+import {
+  Button,
+  Dropdown,
+  Form,
+  Input,
+  List,
+  MenuProps,
+  Modal,
+  Popconfirm,
+  Select,
+  Spin,
+  Typography,
+} from 'antd';
+import { Link, useNavigate } from 'react-router-dom';
+import { PlusOutlined, SortAscendingOutlined } from '@ant-design/icons';
+import { useEffect, useState } from 'react';
+import dayjs from 'dayjs';
+import { saveConfig, config, setPersonListOrder } from 'renderer/ConfigState';
 import { CategoryData, Data, emptyPerson, Person } from './models';
+import styles from './PeopleList.module.css';
 
 const AddPersonModal = ({
   isVisible,
@@ -64,16 +78,88 @@ const AddPersonModal = ({
   );
 };
 
+const EditPersonModal = ({
+  name,
+  isVisible,
+  onCancel,
+  onSave,
+  isSaving,
+}: {
+  name: string;
+  isVisible: boolean;
+  onCancel: () => void;
+  onSave: (name: string) => void;
+  isSaving: boolean;
+}) => {
+  const [newName, setName] = useState(name);
+  const [error, setError] = useState(false);
+  const [form] = Form.useForm();
+  useEffect(() => {
+    // resets to initial value (requires force render or the form thing rest does not work)
+    form.resetFields();
+  }, [name, form]);
+
+  const clear = () => {
+    form.resetFields();
+    setName('');
+    setError(false);
+  };
+  return (
+    <Modal
+      forceRender
+      title="Person Editieren"
+      open={isVisible}
+      confirmLoading={isSaving}
+      onOk={() => {
+        if (newName.trim() === '') setError(true);
+        else {
+          const currName = newName;
+          clear();
+          onSave(currName);
+        }
+      }}
+      onCancel={() => {
+        clear();
+        onCancel();
+      }}
+    >
+      <Form
+        form={form}
+        name="basic"
+        labelCol={{ span: 8 }}
+        wrapperCol={{ span: 16 }}
+        autoComplete="off"
+        initialValues={{ name }}
+      >
+        <Form.Item
+          label="Name"
+          name="name"
+          help={error ? 'Namen eingeben' : undefined}
+          validateStatus={error ? 'error' : 'validating'}
+        >
+          <Input onChange={(val) => setName(val.target.value)} />
+        </Form.Item>
+      </Form>
+    </Modal>
+  );
+};
+
 const EntryDeleteButton = ({ onDelete }: { onDelete: () => Promise<void> }) => {
+  const deleteConfirm = (e: React.MouseEvent | undefined) => {
+    e?.stopPropagation();
+    onDelete();
+  };
   return (
     <Popconfirm
       title="Wollen Sie diese Person wirklich löschen?"
-      onConfirm={onDelete}
-      onCancel={() => {}}
+      onConfirm={deleteConfirm}
+      onCancel={(e) => {
+        e?.stopPropagation();
+      }}
       okText="Weg mit dem!"
       cancelText="Noi doch it!"
     >
-      <Button style={{ padding: '0' }} type="text" size="small">
+      <Button type="text" size="small" onClick={(e) => e.stopPropagation()}>
         Löschen
       </Button>
     </Popconfirm>
@@ -85,15 +171,20 @@ export default function PeopleList({
   categoryData,
   onDeletePerson,
   onAddPerson,
+  onEditPerson,
   isSaving,
 }: {
   data: Data | undefined;
   categoryData: CategoryData | undefined;
   onDeletePerson: (id: string) => Promise<void>;
   onAddPerson: (person: Person) => Promise<void>;
+  onEditPerson: (id: string, name: string) => Promise<void>;
   isSaving: boolean;
 }) {
   const [isNewModalVisible, setIsNewModalVisible] = useState(false);
+  const [editModalId, setEditModalId] = useState<string | undefined>(undefined);
+  const [currentSort, setSort] = useState('name');
+  const navigation = useNavigate();
   if (data == null || categoryData == null) {
     return (
       <div style={{ padding: '40px' }}>
@@ -119,8 +210,34 @@ export default function PeopleList({
     await onAddPerson(newPerson);
     setIsNewModalVisible(false);
   };
+  const handleSortChanged = async (value: string) => {
+    setPersonListOrder(value);
+    await saveConfig();
+    setSort(value);
+  };
+  const sort = [
+    {
+      label: 'Name',
+      value: 'name',
+    },
+    {
+      label: 'Geburtsdatum',
+      value: 'birthday',
+    },
+  ];
 
   const people = Object.entries(data.people);
+  const defaultDate = dayjs('0000-1-01');
+  people.sort((p1, p2) => {
+    if (currentSort === 'name') return p1[1].name.localeCompare(p2[1].name);
+    if (currentSort === 'birthday') {
+      const p1Birthday = dayjs(p1[1].birthday) ?? defaultDate;
+      const p2Birthday = dayjs(p2[1].birthday) ?? defaultDate;
+      return p1Birthday.diff(p2Birthday);
+    }
+
+    return 0;
+  });
   return (
     <>
       <div className="hiddenPrint">
@@ -133,6 +250,15 @@ export default function PeopleList({
           Neue Person hinzufügen
         </Button>
       </div>
+      <div className={styles.sortContainer}>
+        <Select
+          suffixIcon={<SortAscendingOutlined />}
+          style={{ width: '150px' }}
+          defaultValue="name"
+          options={sort}
+          onChange={handleSortChanged}
+        />
+      </div>
       <List
         style={{ backgroundColor: 'white' }}
         bordered
@@ -140,13 +266,27 @@ export default function PeopleList({
         locale={{ emptyText: 'Keine Personen vorhanden' }}
         renderItem={([id, person]: [string, Person]) => (
           <List.Item
+            className={styles.personListItem}
+            onClick={() => navigation(`/person/${id}`)}
             actions={[
-              <EntryDeleteButton onDelete={async () => onDeletePerson(id)} />,
+              <Button
+                type="text"
+                size="small"
+                onClick={(e) => {
+                  e.stopPropagation();
+                  setEditModalId(id);
+                }}
+              >
+                Edit
+              </Button>,
+              <EntryDeleteButton
+                onDelete={async () => {
+                  onDeletePerson(id);
+                }}
+              />,
             ]}
           >
-            <List.Item.Meta
-              title={<Link to={`/person/${id}`}>{person.name}</Link>}
-            />
+            <Typography.Text strong>{person.name}</Typography.Text>
           </List.Item>
         )}
       />
@@ -159,6 +299,20 @@ export default function PeopleList({
         }}
         onSave={(name: string) => {
           handleAddPerson(name);
+        }}
+      />
+      <EditPersonModal
+        name={editModalId != null ? data.people[editModalId].name : ''}
+        isVisible={editModalId != null}
+        isSaving={isSaving}
+        onCancel={() => {
+          setEditModalId(undefined);
+        }}
+        onSave={(name: string) => {
+          if (editModalId != null) {
+            onEditPerson(editModalId, name);
+            setEditModalId(undefined);
+          }
         }}
       />
     </>