+++ src/renderer/index.tsx
@@ -5,6 +5,7 @@ declare module 'valtio' {
   function useSnapshot<T extends object>(p: T): T;
 }
 
+// eslint-disable-next-line @typescript-eslint/no-non-null-assertion
 const container = document.getElementById('root')!;
 const root = createRoot(container);
 root.render(<App />);