+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.7
+
+- Allow renaming of people
+- Sort list by name or birthday
+- Add export as pdf option in print view
+- add area for notes in main view & print view
+- update third-party dependencies
+
 ## 1.6
 
 - Update categories & phases (warning: may contain 'Kot')